pyproject.toml
[project]
name = "argos"
version = "0.1.0"
description = "Local-first professional knowledge graph system"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "python-frontmatter>=1.1.0",
    "python-slugify>=8.0.0",
    "redis>=5.0.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
]

[project.scripts]
kb = "argos.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["argos"]
